Separate out withFeatureToggle HOC component
